requirements.txt
streamlit==1.32.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
fpdf2==2.7.7
Pillow==10.1.0